refactor proposal approvals router
