+++ src/app/routers/proposal_workflow_evidence.py
@@ -1,7 +1,6 @@
 from fastapi import APIRouter, Path
 
 from app.contracts.proposals import (
-    ProposalApprovalsEnvelopeResponse,
     ProposalLineageEnvelopeResponse,
     ProposalWorkflowEventsEnvelopeResponse,
 )
@@ -27,27 +26,6 @@ async def get_workflow_events(
     service = proposal_service()
     correlation_id = correlation_id_var.get()
     return await service.get_workflow_events(
-        proposal_id=proposal_id,
-        correlation_id=correlation_id,
-    )
-
-
-@router.get(
-    "/{proposal_id}/approvals",
-    response_model=ProposalApprovalsEnvelopeResponse,
-    summary="Get Proposal Approvals",
-    description="Returns approval records already captured for a specific advisory proposal.",
-)
-async def get_approvals(
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-) -> ProposalApprovalsEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_approvals(
         proposal_id=proposal_id,
         correlation_id=correlation_id,
     )
